requirements.txt
pandas
xlrd
setuptools>=40.8.0
progressbar2
colorlog

